adding feature to check the user delete & update own posts and prevent to delete posts of other users

+++ controller/feed.js
@@ -1,8 +1,10 @@
-const fs = require('fs')
-const path = require('path')
+const fs = require('fs');
+const path = require('path');
 
-const { validationResult } = require('express-validator')
-const Post = require('../model/post.js')
+const { validationResult } = require('express-validator');
+
+const Post = require('../model/post.js');
+const User = require('../model/user');
 
 const clearImage = filePath => {
     filePath = path.join(__dirname, '..', filePath)
@@ -59,21 +61,28 @@ exports.createPost = (req, res, next) => {
         throw error;
     }
     // creating post class with data from the user
+    let creator;
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        creator: {
-            name: req.body.name
-        },
+        creator: req.userId,
         image: req.file.path.replace("\\" ,"/")
     });
+    
     // saving the post to the database
     post.save().then(postData => {
+        return User.findById(req.userId)
+    }).then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+    }).then(result => {
         res.status(201).json({
         message: "Post Created Succefully",
-        post: postData
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
         })
-    }).catch(error =>  {
+    }).catch(error => {
         if (!error.statusCode) {
             error.statusCode = 500
         }
@@ -92,23 +101,24 @@ exports.updatePost = (req, res, next) => {
     const title = req.body.title
     const content = req.body.content
     let image = req.body.image
-
     if (req.file) {
         image = req.file.path.replace("\\" ,"/")
     }
-    
     if (!image) {
         const error = new Error('No File Picked. !!');
         error.statusCode = 422;
         throw error;
     }
-
     Post.findById(postId).then(post => {
-
         if (!post) {
             const error = new Error('there is no post with this id');
             error.statusCode = 404;
             throw error
+        }
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not Authorized');
+            error.statusCode = 403;
+            throw error;
         }
         if (image !== post.image) {
             clearImage(post.image)
@@ -130,18 +140,25 @@ exports.updatePost = (req, res, next) => {
 
 exports.deletePost = (req, res, next) => {
     const postId = req.params.postId;
-    console.log(postId)
     Post.findById(postId).then(post => {
-        console.log(post)
         if (!post) {
             const error = new Error('There is no post with this id');
             error.statusCode = 422;
             throw error;
         }
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not Authorized');
+            error.statusCode = 403;
+            throw error;
+        }
         clearImage(post.image)
         return Post.findByIdAndRemove(postId)
         }).then(result => {
-            console.log(result)
+            return User.findById(req.userId)
+        }).then(user => {
+            user.posts.pull(postId)
+            return user.save()
+        }).then(result => {
             res.status(200).json({message: 'Post Deleted !!!'})
         }).catch(error =>  {
             if (!error.statusCode) {
